+++ controllers/orderController.js
@@ -1,13 +1,7 @@
 const orderService = require("../services/orderService");
 const { asyncErrorHandler } = require("../middlewares/errorHandling");
 
-// (1) 사용자 정보 리턴 (BE -> FE)
-const getUserInfo = asyncErrorHandler(async (req, res) => {
-  const userInfo = await orderService.getUserInfo(req.userId);
-  return res.status(200).json(userInfo);
-});
-
-// (2) 주문/결제 && 주문번호(id) 리턴
+// 주문/결제 && 주문번호(id) 리턴
 const order = asyncErrorHandler(async (req, res) => {
   const { toName, toMobile, toAddress } = req.body;
 
@@ -25,7 +19,7 @@ const order = asyncErrorHandler(async (req, res) => {
   return res.status(200).json({ orderId: orderId });
 });
 
-// (3) 완료된 주문 조회
+// 완료된 주문 조회
 const getOrderResult = asyncErrorHandler(async (req, res) => {
   const { orderId } = req.params;
 
@@ -39,7 +33,6 @@ const getOrderResult = asyncErrorHandler(async (req, res) => {
 });
 
 module.exports = {
-  getUserInfo,
   order,
   getOrderResult,
 };
